+++ src/lib/common/FlexContainer/FlexContainer.stories.tsx
@@ -7,7 +7,7 @@ import { sizeArgTypes } from '../styles/size';
 import { spacingArgTypes } from '../styles/spacing';
 import type { FlexContainerProps } from './FlexContainer';
 import FlexContainer from './FlexContainer';
-import { LAYOUT } from './style';
+import { LAYOUT } from './styles/flexContainer.style';
 
 const generateChildrenBoxes = () => {
   const BOX_COUNT = 6;
