+++ src/lib/common/FlexContainer/FlexContainer.tsx
@@ -2,8 +2,8 @@ import styled from '@emotion/styled';
 
 import type { ContainerProps } from '../Container/Container';
 import Container from '../Container/Container';
-import type { Gap, Layout } from './style';
-import { gapStyle, layoutStyle } from './style';
+import type { Gap, Layout } from './styles/flexContainer.style';
+import { gapStyle, layoutStyle } from './styles/flexContainer.style';
 
 export type BasicLayoutProps = Gap & ContainerProps;
