+++ src/lib/common/Tab/Menus.tsx
@@ -1,11 +1,10 @@
-import { css } from '@emotion/react';
-
 import type { PropsWithChildren } from 'react';
 import { useContext } from 'react';
 
 import FlexContainer from '../FlexContainer';
 import type { FlexContainerProps } from '../FlexContainer/FlexContainer';
 import { TabContext } from './Tab';
+import { menuContainerStyle } from './styles/menus.style';
 
 export interface MenusProps extends FlexContainerProps {
   /** 선택된 탭 메뉴를 강조하는 색깔 변경 가능
@@ -36,36 +35,12 @@ const Menus = ({
       justify={vertical ? 'start' : 'space-between'}
       position="left"
       gap={gap}
-      css={menuStyle({ highlightColor, noUnderline, vertical })}
+      css={menuContainerStyle({ highlightColor, noUnderline, vertical })}
       {...attributes}
     >
       {children}
     </FlexContainer>
   );
 };
-
-interface menuStyleProps {
-  highlightColor: string;
-  noUnderline: boolean;
-  vertical: boolean | undefined;
-}
-
-const menuStyle = ({ highlightColor, noUnderline, vertical = false }: menuStyleProps) => css`
-  ${!noUnderline && `box-shadow: 0 2px 0 0 #eee;`}
-
-  button {
-    ${vertical && `box-shadow: 0 2px 0 0 #eee;`}
-  }
-
-  button.active {
-    font-weight: 600;
-    color: ${highlightColor};
-
-    ${!noUnderline && `box-shadow: 0 2px 0 0 ${highlightColor};`}
-
-    fill: ${highlightColor};
-    stroke: ${highlightColor};
-  }
-`;
 
 export default Menus;
