+++ src/lib/common/Tab/styles/menu.style.ts
@@ -1,0 +1,40 @@
+import { css } from '@emotion/react';
+
+import type { IconPosition } from '../Menu';
+
+export const tabMenuStyle = css`
+  flex: 1;
+
+  padding: 8px 0;
+
+  color: #555;
+  text-align: center;
+
+  background: transparent;
+  border: none;
+
+  &:hover {
+    color: #333;
+  }
+`;
+
+export const getFlexDirection = (iconPosition: IconPosition) => {
+  switch (iconPosition) {
+    case 'top':
+      return 'column';
+    case 'bottom':
+      return 'column-reverse';
+    case 'left':
+      return 'row';
+    case 'right':
+      return 'row-reverse';
+  }
+};
+
+export const tabMenuWithIconStyle = (iconPosition: IconPosition) => css`
+  display: flex;
+  flex-direction: ${getFlexDirection(iconPosition)};
+  gap: 2px;
+  align-items: center;
+  justify-content: center;
+`;
